Moved app_func to define_probe_function and updated api file to import everything. Removed unneeded __init__ file. Deleted Circle_overlap since it did not contain proper function
