+++ epsic_tools/api.py
@@ -6,3 +6,6 @@ Created on Fri Feb 14 14:46:57 2020
 """
 
 from epsic_tools import ptychography2D
+from epsic_tools.toolbox import define_probe_function
+from epsic_tools.toolbox import radial_profile
+from epsic_tools.toolbox import warp_3d
