+++ epsic_tools/toolbox/.gitignore
@@ -1,1 +1,2 @@
-__pycashe__/
+__pycache__/
+ptychography/__pycache__/
